LPS-192184 Move Bind and Unbind tests to ObjectDefinitionLocalServiceTest
